Day 6 optimize further

+++ day06/benches/bench.rs
@@ -3,9 +3,9 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use day06::*;
 
 fn criterion_benchmark(c: &mut Criterion) {
-    let data: Vec<u8> = include_str!("../input.txt")
+    let data: Vec<usize> = include_str!("../input.txt")
         .split(',')
-        .map(|n| n.parse::<u8>().unwrap())
+        .map(|n| n.parse::<usize>().unwrap())
         .collect();
 
     c.bench_function("day06 task_1", |b| b.iter(|| task_1(black_box(&data))));
