+++ day06/src/main.rs
@@ -2,7 +2,10 @@ use day06::*;
 
 fn main() {
     let input = include_str!("../input.txt");
-    let data: Vec<u8> = input.split(',').map(|n| n.parse::<u8>().unwrap()).collect();
+    let data: Vec<usize> = input
+        .split(',')
+        .map(|n| n.parse::<usize>().unwrap())
+        .collect();
 
     println!("Task 1: {}", task_1(&data));
     println!("Task 2: {}", task_2(&data));
